feat: add method for enqueuing a listener before others
By default, custom key listeners execute in the same order they were registered.
In some cases, it's desirable to add a listener that executes before others, so that it can prevent the event from propagating, so that the event is not handled by other listeners that have been added earlier.